# docs/architecture/openapi.yaml
openapi: 3.0.3
info:
  title: Habitree Backend API
  version: 1.1.0
  description: Aktuelle Definition der REST-Endpunkte des Habitree-Backends (Express + Prisma + JWT).
servers:
  - url: http://iseproject01.informatik.htw-dresden.de:8000
    description: Externer Server
  - url: http://localhost:8000
    description: Lokale Entwicklung
paths:
...
  /user/status:
    get:
      summary: Login-Status des aktuellen Nutzers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Status des eingeloggten Nutzers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStatusResponse'
        '401': { description: Nicht autorisiert }
  /user/streak:
    get:
      summary: Streak-Informationen des aktuellen Nutzers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Streak-Daten
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStreakResponse'
        '401': { description: Nicht autorisiert }
  /user/username:
    put:
      summary: Benutzernamen aktualisieren
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUsernameRequest'
      responses:
        '200':
          description: Benutzername geändert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateUsernameResponse'
        '400': { description: Ungültige Eingabe }
        '401': { description: Nicht autorisiert }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthUser:
      type: object
      properties:
        userId: { type: integer }
        email: { type: string, format: email }
        username: { type: string }
    UserStatusResponse:
      type: object
      properties:
        isLoggedIn: { type: boolean }
        userId: { type: integer }
        email: { type: string, format: email }
        username: { type: string }
    UserStreakResponse:
      type: object
      properties:
        userId: { type: integer }
        currentStreak: { type: integer }
        maxStreak: { type: integer }
        lastCompletionDate:
          type: string
          format: date-time
    UpdateUsernameRequest:
      type: object
      required: [username]
      properties:
        username:
          type: string
          minLength: 3
    UpdateUsernameResponse:
      type: object
      properties:
        message: { type: string }
        user:
          $ref: '#/components/schemas/AuthUser'
...